Improve readability of the release notes
Without this patch, one user could read the release note, and
completely not know that manual intervention need to happen,
nor its high level impact.

This clarifies it by refering to the upgrade notes from the
release note of the version. The upgrades notes are then sorted
in the chronological order of things that need to happen.