Replace control plane term with master node
Without this patch, it could be unclear what "control plane"
node means. "kubelet" could technically be considered as part
of the control plane if we think a little outside the box.

Let's make this clear. We are talking about "nodes" here, and the
"master" nodes come before the "worker" nodes.