Remove reference to passwords
password was a variable introduced in beta, for convenience, and
was never documented or recommended.
Configuring password might trigger a user to define a password,
even though it's not necessary. This is naturally not recommended.

Without this patch, the user might want to define new fields,
which shouldn't be defined. This is a problem for security purposes.

This fixes it by removing the recommendation to define password.
If a pre-beta user is trying to use password, he will get a
warning, and re-evaluate. This is for the best.